fix: preserve sender identity on received email replies
