+++ frontend/src/components/Activities/EmailArea.vue
@@ -93,7 +93,9 @@ function reply(email, reply_all = false) {
   let editor = emailBox.editor
   let message = email.content
   let recipients = email.recipients.split(',').map((r) => r.trim())
-  editor.fromEmail = email.sender
+  if (email.sent_or_received === 'Sent') {
+    editor.fromEmail = email.sender
+  }
   editor.toEmails = [email.sender]
   editor.cc = editor.bcc = false
   editor.ccEmails = []
